Fix params build function
